WPI (3/11/17) Post-WPI
Lots of changes, including:
- increasing the error for the encoders, which may either may need to be
increased to ensure success, or dialed in to keep our autonomous codes
as sucessful as they were at WPI
-Changing most of the shooter code and removing our CountRPM class due
to the Nullpointer it was giving. The class itself was not considered a
PIDSource which we had done by having the class implement PIDSource.
Needs to be fixed before Bryant
-Changing the Climbing of our Robot to have  up and down motions without
the use of the limit switches. If we see fit again to do so, we will
reimplement the Limit Switches to our code, however that depends on
whether we will have a Limit Switch with our new climber