Beginning of The Thing :camel:  :  Gets Angle, TargetCenter and DistanceFromCenter. Not Usable  Yet.
Just as a percursor: Thanks to Mr. Elmhurst and Stewie for setting up
the rough code and thinking of using the PixyCam to solve most of the
processing problems.

This stuff that I have put together for the time being. It is nowhere
near done, but I am Comitting this now so that I don't accidently lose
this through some magic means. I've tried my hand at documenting to the
best of my ability the commands that I've created. If anyone wants to
edit this, I hope the Documentation helps.